# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WatchlistEntry {
  id: number;
  username: string;
  ticker: string;
  addedAt: string;
}

export interface AddToWatchlistBody {
  ticker: string;
  username?: string;
}

export interface StockMetrics {
  ticker: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  peRatioForward?: number | null;
  /** @nullable */
  epsGrowthYoy?: number | null;
  /** @nullable */
  debtToEquity?: number | null;
  /** @nullable */
  ma200?: number | null;
  /** @nullable */
  ma50?: number | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  shortInterestPct?: number | null;
  /** @nullable */
  putCallRatio?: number | null;
  /** @nullable */
  beta?: number | null;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface HistoricalDataPoint {
  date: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  ma50?: number | null;
  /** @nullable */
  ma200?: number | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  volume?: number | null;
}

export interface StockHistory {
  ticker: string;
  period: string;
  dataPoints: HistoricalDataPoint[];
}

export interface ErrorResponse {
  error: string;
}

export type GetWatchlistParams = {
  username?: string;
};

export type GetStockHistoryParams = {
  period?: GetStockHistoryPeriod;
};

export type GetStockHistoryPeriod =
  (typeof GetStockHistoryPeriod)[keyof typeof GetStockHistoryPeriod];

export const GetStockHistoryPeriod = {
  "1mo": "1mo",
  "3mo": "3mo",
  "6mo": "6mo",
  "1y": "1y",
  "2y": "2y",
} as const;

export type GetWatchlistMetricsParams = {
  username?: string;
};