# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the list of stock tickers in the user's watchlist
 * @summary Get user watchlist
 */
export const getWatchlistQueryUsernameDefault = `demo`;

export const GetWatchlistQueryParams = zod.object({
  username: zod.coerce.string().default(getWatchlistQueryUsernameDefault),
});

export const GetWatchlistResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  ticker: zod.string(),
  addedAt: zod.coerce.date(),
});
export const GetWatchlistResponse = zod.array(GetWatchlistResponseItem);

/**
 * @summary Add a ticker to the watchlist
 */
export const AddToWatchlistBody = zod.object({
  ticker: zod.string(),
  username: zod.string().optional(),
});

/**
 * @summary Remove a ticker from the watchlist
 */
export const RemoveFromWatchlistParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get current metrics for a stock ticker
 */
export const GetStockMetricsParams = zod.object({
  ticker: zod.coerce.string(),
});

export const GetStockMetricsResponse = zod.object({
  ticker: zod.string(),
  companyName: zod.string().nullish(),
  currentPrice: zod.number().nullish(),
  peRatioForward: zod.number().nullish(),
  epsGrowthYoy: zod.number().nullish(),
  debtToEquity: zod.number().nullish(),
  ma200: zod.number().nullish(),
  ma50: zod.number().nullish(),
  rsi: zod.number().nullish(),
  shortInterestPct: zod.number().nullish(),
  putCallRatio: zod.number().nullish(),
  beta: zod.number().nullish(),
  lastUpdated: zod.coerce.date().nullish(),
});

/**
 * @summary Get historical metric data for charting
 */
export const GetStockHistoryParams = zod.object({
  ticker: zod.coerce.string(),
});

export const getStockHistoryQueryPeriodDefault = `6mo`;

export const GetStockHistoryQueryParams = zod.object({
  period: zod
    .enum(["1mo", "3mo", "6mo", "1y", "2y"])
    .default(getStockHistoryQueryPeriodDefault),
});

export const GetStockHistoryResponse = zod.object({
  ticker: zod.string(),
  period: zod.string(),
  dataPoints: zod.array(
    zod.object({
      date: zod.string(),
      price: zod.number().nullish(),
      ma50: zod.number().nullish(),
      ma200: zod.number().nullish(),
      rsi: zod.number().nullish(),
      volume: zod.number().nullish(),
    }),
  ),
});

/**
 * @summary Get current metrics for all tickers in the watchlist
 */
export const getWatchlistMetricsQueryUsernameDefault = `demo`;

export const GetWatchlistMetricsQueryParams = zod.object({
  username: zod.coerce
    .string()
    .default(getWatchlistMetricsQueryUsernameDefault),
});

export const GetWatchlistMetricsResponseItem = zod.object({
  ticker: zod.string(),
  companyName: zod.string().nullish(),
  currentPrice: zod.number().nullish(),
  peRatioForward: zod.number().nullish(),
  epsGrowthYoy: zod.number().nullish(),
  debtToEquity: zod.number().nullish(),
  ma200: zod.number().nullish(),
  ma50: zod.number().nullish(),
  rsi: zod.number().nullish(),
  shortInterestPct: zod.number().nullish(),
  putCallRatio: zod.number().nullish(),
  beta: zod.number().nullish(),
  lastUpdated: zod.coerce.date().nullish(),
});
export const GetWatchlistMetricsResponse = zod.array(
  GetWatchlistMetricsResponseItem,
);