# lib/api-zod/src/generated/types/addToWatchlistBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */

export interface AddToWatchlistBody {
  ticker: string;
  username?: string;
}