# lib/api-zod/src/generated/types/historicalDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoricalDataPoint {
  date: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  ma50?: number | null;
  /** @nullable */
  ma200?: number | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  volume?: number | null;
}