# lib/api-zod/src/generated/types/stockHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoricalDataPoint } from "./historicalDataPoint";

export interface StockHistory {
  ticker: string;
  period: string;
  dataPoints: HistoricalDataPoint[];
}