# lib/api-zod/src/generated/types/stockMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */

export interface StockMetrics {
  ticker: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  peRatioForward?: number | null;
  /** @nullable */
  epsGrowthYoy?: number | null;
  /** @nullable */
  debtToEquity?: number | null;
  /** @nullable */
  ma200?: number | null;
  /** @nullable */
  ma50?: number | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  shortInterestPct?: number | null;
  /** @nullable */
  putCallRatio?: number | null;
  /** @nullable */
  beta?: number | null;
  /** @nullable */
  lastUpdated?: Date | null;
}