# lib/api-zod/src/generated/types/watchlistEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Watchlist Supercharger API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistEntry {
  id: number;
  username: string;
  ticker: string;
  addedAt: Date;
}